[Robot] Set default arm motion profile based on testing
These were about the best values that I achieved in testing.

kP was much smoother at 0.02. Anything higher resulted in shakey movement while the arm was swinging

Acceleration seemed to play the biggest role in not tipping the robot so I increased it until the robot tipped a bit while coming back from cube high

Velocity could be increased pretty high but having it too high caused tippiness when coming back from cube high because it was moving all of the mass at the end of the arm really quickly, causing a backwards tip